Remove KotlinPoet and old generator
