+++ build.gradle
@@ -25,7 +25,6 @@ buildscript {
         daggerVersion = '2.15'
         kotlinxCoroutinesVersion = '0.23.3'
         kotlinxCollectionsImmutableVersion = '0.1'
-        kotlinPoetVersion='1.0.0-RC1'
     }
 
     repositories {
