+++ modules/core/arrow-annotations-processor/build.gradle
@@ -5,7 +5,6 @@ apply plugin: 'kotlin-kapt'
 
 dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlinVersion"
-    compile "com.squareup:kotlinpoet:$kotlinPoetVersion"
     compile project(':arrow-annotations')
     compile 'me.eugeniomarletti.kotlin.metadata:kotlin-metadata:1.4.0'
     compileOnly 'com.google.auto.service:auto-service:1.0-rc4'
